Add module option tab completion to run command
