+++ lib/msf/ui/console/command_dispatcher.rb
@@ -135,5 +135,5 @@ end
 end end end
 
 require 'msf/ui/console/module_command_dispatcher'
+require 'msf/ui/console/module_option_tab_completion'
 require 'msf/ui/console/command_dispatcher/core'
-
