+++ lib/msf/ui/console/command_dispatcher/exploit.rb
@@ -12,6 +12,7 @@ module CommandDispatcher
 class Exploit
 
   include Msf::Ui::Console::ModuleCommandDispatcher
+  include Msf::Ui::Console::ModuleOptionTabCompletion
 
   @@exploit_opts = Rex::Parser::Arguments.new(
     "-e" => [ true,  "The payload encoder to use.  If none is specified, ENCODER is used." ],
@@ -69,21 +70,31 @@ class Exploit
     return session
   end
 
-  def cmd_exploit_tabs(str, words)
+  #
+  # Tab completion for the run command
+  #
+  def cmd_run_tabs(str, words)
     fmt = {
-      '-e' => [ framework.encoders.map { |refname, mod| refname } ],
-      '-f' => [ nil                                               ],
-      '-h' => [ nil                                               ],
-      '-j' => [ nil                                               ],
-      '-J' => [ nil                                               ],
-      '-n' => [ framework.nops.map { |refname, mod| refname }     ],
-      '-o' => [ true                                              ],
-      '-p' => [ framework.payloads.map { |refname, mod| refname } ],
-      '-t' => [ true                                              ],
-      '-z' => [ nil                                               ]
+        '-e' => [ framework.encoders.map { |refname, mod| refname } ],
+        '-f' => [ nil                                               ],
+        '-h' => [ nil                                               ],
+        '-j' => [ nil                                               ],
+        '-J' => [ nil                                               ],
+        '-n' => [ framework.nops.map { |refname, mod| refname }     ],
+        '-o' => [ true                                              ],
+        '-p' => [ framework.payloads.map { |refname, mod| refname } ],
+        '-t' => [ true                                              ],
+        '-z' => [ nil                                               ]
     }
-    tab_complete_generic(fmt, str, words)
+    flags = tab_complete_generic(fmt, str, words)
+    options = tab_complete_option(str, words)
+    flags + options
   end
+
+  #
+  # Tab completion for the exploit command
+  #
+  alias cmd_exploit_tabs cmd_run_tabs
 
   #
   # Launches exploitation attempts.
